+++ resources/files/documentFileUpload.py
@@ -2,7 +2,7 @@ from flask_restful import reqparse, Resource
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from werkzeug.datastructures import FileStorage
 from werkzeug import Response
-from Injectors import FilesContainer
+from dependencyInjectors import FilesContainer
 from enums import ErrorEnums
