+++ service/userLoginService.py
@@ -1,4 +1,4 @@
-import dto.UserDTO
+import dataStateAccess.UserDTO
 from databaseService.userDatabaseService import UserDatabaseService
 import utils.SecurityUtils as EncryptPass
 
@@ -7,7 +7,7 @@ def get_by_username(username):
     user = UserDatabaseService.get_user_by_username(username)
     if user and not user.is_active and user.marked_active_inactive_by_admin:
         return "User is currently inactive. Please contact admin."
-    user_dto = dto.UserDTO.user_dto(user)
+    user_dto = dataStateAccess.UserDTO.user_dto(user)
     return user_dto
 
 
@@ -17,13 +17,13 @@ def get_by_email(email):
         return "No user found with email - {}.".format(email)
     if user and not user.is_active and user.marked_active_inactive_by_admin:
         return "User is currently inactive. Please contact admin"
-    user_dto = dto.UserDTO.user_dto(user)
+    user_dto = dataStateAccess.UserDTO.user_dto(user)
     return user_dto
 
 
 def get_active_inactive_users_by_email(email):
     user_hold = UserDatabaseService.get_active_inactive_single_user_by_email(email)
-    return dto.UserDTO.user_dto(user_hold)
+    return dataStateAccess.UserDTO.user_dto(user_hold)
 
 
 def convert_password(p) -> str:
